+++ SpringMVC_Security_Example-master/src/main/java/jm/security/example/dao/UserDao.java
@@ -15,11 +15,11 @@ public interface UserDao {
 
     void updateUser(User user);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     User getUserByName(String name);
 
     List<Role> allUsersRole();
 
-    Role getRoleById(int id);
+    Role getRoleById(Long id);
 }
